refactor: move diabetesdbfilteritem pro model do back e remove types.ts
