+++ src/features/acf/diabetes/frontend/AcfPage/AcfPage.tsx
@@ -7,7 +7,7 @@ import type { Session } from "next-auth";
 import { getServerSession } from "next-auth";
 import { ErrorPage } from "./modules/ErrorPage";
 import { PanelSelectorContainer } from "./modules/PanelSelector/PanelSelector.container";
-import type { AcfDashboardType } from "../../common/types";
+import type { AcfDashboardType } from "../../common/model";
 
 export const AcfDashboardPage = async ({
     searchParams,
