+++ src/features/acf/diabetes/frontend/AcfPage/modules/PanelSelector/modules/dashboards/modules/cards/externalCards/externalCards.adapter.ts
@@ -1,4 +1,4 @@
-import type { AcfDashboardType } from "../../../../../../../../../common/types"
+import type { AcfDashboardType } from "../../../../../../../../../common/model"
 import type { DiabetesAcfExternalCardsDescriptionType, ExternalCardDBDataItem, ExternalCardDataItem } from "./externalCards.model"
